requirements.txt
asgiref==3.5.2
black==22.8.0     
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
dj-rest-auth==2.2.5
Django==4.1.1
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
drf-flex-fields==1.0.0
idna==3.4
mypy-extensions==0.4.3